mode useAdvancedRender to dream block renderer
